Expose `appendChild`
This way yo-yoify can reuse the exact same behaviour `bel` has, by
outputting `require('bel/appendChild')`.
